requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AI classification (optional)
google-generativeai>=0.3.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AI classification (optional)
google-generativeai>=0.3.0

# Google Sheets integration (optional)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

requests
beautifulsoup4
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
streamlit
tqdm
fastapi
uvicorn

# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# HTTP and web scraping
requests==2.31.0
jinja2==3.1.3

# Playwright for dynamic JS rendering (optional but recommended)
playwright==1.40.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Image processing (for color extraction)
Pillow==10.2.0
scikit-learn==1.4.0

# Google Sheets integration (optional)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# AI integration (optional)
# google-generativeai==0.3.2  # Uncomment if using Gemini API